nuevo splash screen con circular progress bar

+++ GamaPFC/Gama.Socios.DataAccess/NHibernateSessionFactory.cs
@@ -31,9 +31,17 @@ namespace Gama.Socios.DataAccess
                     try
                     {
                         NHibernate.Cfg.Configuration configuration;
-                        var path = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + @"\nh_socios.cfg";
 
-                        File.Delete(path);
+                        var directoryPath = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)
+                                + @"\GamaData\";
+
+                        if (!Directory.Exists(directoryPath))
+                            Directory.CreateDirectory(directoryPath);
+
+                        var path = directoryPath + @"\nh_socios.cfg";
+
+                        if (File.Exists(path)) { File.Delete(path); }
+
                         if (File.Exists(path))
                         {
                             var file = File.Open(path, FileMode.Open);
@@ -83,7 +91,7 @@ namespace Gama.Socios.DataAccess
                         c.SetProperty("current_session_context_class", "thread_static");
                         schema.Execute(
                             useStdOut: false,
-                            execute: false,
+                            execute: true,
                             justDrop: false);
                     })
                 .BuildConfiguration();
